+++ lib/multi_git/config.rb
@@ -1,0 +1,15 @@
+module MultiGit
+  module Config
+
+    DEFAULTS = {
+      'core.logallrefupdates' => 'false'
+    }
+
+    include Enumerable
+
+    def to_h
+      Hash[to_a]
+    end
+
+  end
+end