+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.1.beta1
 
+* [FEATURE] Repository#config
 * [ENHANCEMENT] Tree#glob now supports DOTMATCH
 * [BUG] Removed Tree#path ( did not made much sense since the path is not known )
 * [BUG] Detaching symbolic references now works as expected